fix(route): 哈尔滨工业大学（威海）通知公告 (#10801)
* Update today.js

* Update today.js

* refactor: migrate to v2
