+++ lib/router.js
@@ -535,7 +535,7 @@ router.get('/hhu/libNewsc', lazyloadRouteHandler('./routes/universities/hhu/libN
 router.get('/hitsz/article/:category?', lazyloadRouteHandler('./routes/universities/hitsz/article'));
 
 // 哈尔滨工业大学（威海）
-router.get('/hitwh/today', lazyloadRouteHandler('./routes/universities/hitwh/today'));
+// router.get('/hitwh/today', lazyloadRouteHandler('./routes/universities/hitwh/today'));
 
 // 上海科技大学
 router.get('/shanghaitech/activity', lazyloadRouteHandler('./routes/universities/shanghaitech/activity'));
