+++ lib/v2/hitwh/maintainer.js
@@ -1,0 +1,3 @@
+module.exports = {
+    '/today': ['raptazure'],
+};
