+++ lib/v2/hitwh/radar.js
@@ -1,0 +1,13 @@
+module.exports = {
+    'hitwh.edu.cn': {
+        _name: '哈尔滨工业大学（威海）',
+        '.': [
+            {
+                title: '今日工大 - 通知公告',
+                docs: 'https://docs.rsshub.app/university.html#ha-er-bin-gong-ye-da-xue-wei-hai',
+                source: ['/1024/list.htm', '/'],
+                target: '/hitwh/today',
+            },
+        ],
+    },
+};
