+++ lib/v2/hitwh/router.js
@@ -1,0 +1,3 @@
+module.exports = (router) => {
+    router.get('/today', require('./today'));
+};
